+++ lib/components/card/card.dart
@@ -1,0 +1,1 @@
+export 'new_plan_card.dart';
