+++ lib/components/components.dart
@@ -2,3 +2,4 @@ export 'dialog/alert_dialog.dart';
 export 'input/input.dart';
 export 'others/others.dart';
 export 'dialog/alert_plan.dart';
+export 'card/card.dart';
